Move Broker Message generation to separate module (outside orm resource)
