+++ nucliadb/nucliadb/ingest/orm/resource.py
@@ -383,102 +383,6 @@ class Resource:
                     )
         return brain
 
-    async def generate_field_vectors(
-        self,
-        bm: BrokerMessage,
-        type_id: FieldType.ValueType,
-        field_id: str,
-        field: Field,
-    ):
-        vo = await field.get_vectors()
-        if vo is None:
-            return
-        evw = ExtractedVectorsWrapper()
-        evw.field.field = field_id
-        evw.field.field_type = type_id  # type: ignore
-        evw.vectors.CopyFrom(vo)
-        bm.field_vectors.append(evw)
-
-    async def generate_field_large_computed_metadata(
-        self,
-        bm: BrokerMessage,
-        type_id: FieldType.ValueType,
-        field_id: str,
-        field: Field,
-    ):
-        lcm = await field.get_large_field_metadata()
-        if lcm is None:
-            return
-        lcmw = LargeComputedMetadataWrapper()
-        lcmw.field.field = field_id
-        lcmw.field.field_type = type_id  # type: ignore
-        lcmw.real.CopyFrom(lcm)
-        bm.field_large_metadata.append(lcmw)
-
-    async def generate_field_computed_metadata(
-        self,
-        bm: BrokerMessage,
-        type_id: FieldType.ValueType,
-        field_id: str,
-        field: Field,
-    ):
-        fcmw = FieldComputedMetadataWrapper()
-        fcmw.field.field = field_id
-        fcmw.field.field_type = type_id  # type: ignore
-
-        field_metadata = await field.get_field_metadata()
-        if field_metadata is not None:
-            fcmw.metadata.CopyFrom(field_metadata)
-            fcmw.field.field = field_id
-            fcmw.field.field_type = type_id  # type: ignore
-            bm.field_metadata.append(fcmw)
-            # Make sure cloud files are removed for exporting
-
-    async def generate_extracted_text(
-        self,
-        bm: BrokerMessage,
-        type_id: FieldType.ValueType,
-        field_id: str,
-        field: Field,
-    ):
-        etw = ExtractedTextWrapper()
-        etw.field.field = field_id
-        etw.field.field_type = type_id  # type: ignore
-        extracted_text = await field.get_extracted_text()
-        if extracted_text is not None:
-            etw.body.CopyFrom(extracted_text)
-            bm.extracted_text.append(etw)
-
-    async def generate_field(
-        self,
-        bm: BrokerMessage,
-        type_id: FieldType.ValueType,
-        field_id: str,
-        field: Field,
-    ):
-        # Used for exporting a field
-        if type_id == FieldType.TEXT:
-            value = await field.get_value()
-            bm.texts[field_id].CopyFrom(value)
-        elif type_id == FieldType.LINK:
-            value = await field.get_value()
-            bm.links[field_id].CopyFrom(value)
-        elif type_id == FieldType.FILE:
-            value = await field.get_value()
-            bm.files[field_id].CopyFrom(value)
-        elif type_id == FieldType.CONVERSATION:
-            value = await self.get_full_conversation(field)  # type: ignore
-            bm.conversations[field_id].CopyFrom(value)
-        elif type_id == FieldType.KEYWORDSET:
-            value = await field.get_value()
-            bm.keywordsets[field_id].CopyFrom(value)
-        elif type_id == FieldType.DATETIME:
-            value = await field.get_value()
-            bm.datetimes[field_id].CopyFrom(value)
-        elif type_id == FieldType.LAYOUT:
-            value = await field.get_value()
-            bm.layouts[field_id].CopyFrom(value)
-
     async def get_full_conversation(
         self,
         conversation_field: Conversation,
@@ -496,56 +400,6 @@ class Resource:
             full_conv.messages.extend(page.messages)
             n_page += 1
         return full_conv
-
-    async def generate_broker_message(self) -> BrokerMessage:
-        # full means downloading all the pointers
-        # minuts the ones to external files that are not PB
-        # Go for all fields and recreate brain
-        bm = BrokerMessage()
-        bm.kbid = self.kb.kbid
-        bm.uuid = self.uuid
-        basic = await self.get_basic()
-        if basic is not None:
-            bm.basic.CopyFrom(basic)
-        bm.slug = bm.basic.slug
-        origin = await self.get_origin()
-        if origin is not None:
-            bm.origin.CopyFrom(origin)
-        relations = await self.get_relations()
-        if relations is not None:
-            for relation in relations.relations:
-                bm.relations.append(relation)
-
-        fields = await self.get_fields(force=True)
-        for (type_id, field_id), field in fields.items():
-            # Value
-            await self.generate_field(bm, type_id, field_id, field)
-
-            # Extracted text
-            await self.generate_extracted_text(bm, type_id, field_id, field)
-
-            # Field Computed Metadata
-            await self.generate_field_computed_metadata(bm, type_id, field_id, field)
-
-            if type_id == FieldType.FILE and isinstance(field, File):
-                field_extracted_data = await field.get_file_extracted_data()
-                if field_extracted_data is not None:
-                    bm.file_extracted_data.append(field_extracted_data)
-
-            elif type_id == FieldType.LINK and isinstance(field, Link):
-                link_extracted_data = await field.get_link_extracted_data()
-                if link_extracted_data is not None:
-                    bm.link_extracted_data.append(link_extracted_data)
-
-            # Field vectors
-            await self.generate_field_vectors(bm, type_id, field_id, field)
-
-            # Large metadata
-            await self.generate_field_large_computed_metadata(
-                bm, type_id, field_id, field
-            )
-
-        return bm
 
     # Fields
     async def get_fields(
